improve: separate typings in subfolders for more organization
this helps making it cleaner for identifying the types.
