+++ src/managers/trackCacheManager.ts
@@ -1,6 +1,6 @@
 import crypto from 'node:crypto'
 import fs from 'node:fs/promises'
-import type { TrackCacheEntry } from '../typings/trackCache.types.ts'
+import type { TrackCacheEntry } from '../typings/playback/trackCache.types.ts'
 import { logger } from '../utils.js'
 
 const TRACK_CACHE_SALT = 'nodelink-track-salt'
