+++ src/typings/voice/connection.types.ts
